rename filter.py to query_filters.py
